Rename servlet common snippet package
