+++ instrumentation/servlet/servlet-common/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/servlet/common/snippet/OutputStreamSnippetInjectionHelper.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.servlet.snippet;
+package io.opentelemetry.javaagent.instrumentation.servlet.common.snippet;
 
 import static java.util.logging.Level.FINE;
